Add Basic Show Page For Products
This adds a very simple link from the home page and a simple show page
for the product. I added a price range calculator in order to support
showing the range of prices from the listing page based upon skus. It
essentially takes the upper end and lower end and joins them. If only 1
price is present the join doesn't happen and it just surfaces the price.