Build out Order Items Flow
When a user adds an item to the cart it will be through an order item
form. This form contains the sku_id and will then send to create an
order_item with that sku_id, the default quantity of 1 and then return
to the products page.

These order items will eventually get tied to an order through the user
of a session id or a user id.