Add Order Model
This creates the order model. It is pretty simple since it is mostly a
container for order_items. It is mandated to have a session id which
will be tied to the users cookie session to persist until they clear
cookies. Once a user signs in then we can associate any outstanding
orders to the user with the session id.