Create Order with Order Items
An Order is a container for a shopping cart. When a user first places an
item in the cart it attempts to find an order by their session id and a
non abandoned shopping cart. If found it adds the order item to that, if
not found then it first creates an order and then creates the order
item.