Set Up Shipping Form Model
This takes the shipping form as far as I can go without the shipping
address model. Shipping Form is a form object that wraps the creation of
the shipping address with the creation of a user if they supply a
password. It will then create the shipping address and tie the shipping
address to the user.