Checkout Workflow
This is a high level very basic implementation of the checkout workflow.
Current_order becomes a key method that drives much of the checkout. The
overview is they hit a page to review the cart, proceed to enter their
shipping info, then billing info, then review and hit process to charge
the card.

On the shipping info step a password is asked for. The intent is to
either log them into their account or create an account if one doesn't
exist. This functionality does not yet exist. If the user does in fact
have an account or enter a password to create an account, the credit
card should get tied to the user and a user created in stripe. This
functionality is stubbed but not implemented.

Otherwise the token is used as a one time charge and the funding
instrument can no longer be used.