Finish Users That Sign Up Check Out Flow
This completes the scaffold for users that sign up by entering a
password on the shipping form. There is some ugliness that gets
introduced that is good candidates for refactoring later.

Devise creates a new session_id when the user logs in. Actually warden
does it and there is a flag to disable with (:renew) but that's not good
from a security stand point. Instead I decided to update the carts when
the user logs in and then search primarily by the user_id with a
fallback to session id if there is no user.

I decided to use the create action as the ideal place to show the
receipt. At that point the user can print or navigate away from the
page but won't be able to return from there for now.