Header Styling and Cart Sidebar
This changes the header to have a left and right section. It also
introduces the cart sidebar that will pop over everything else. Styles
are not flushed out at this point nor is content but the functionality
is at a rough first page stage. Using components from refills to
construct the menu and sidebars.