Create Image Upload for Products
This implements the image upload feature for products. It utilizes
cloudinary underneath to handle all the resizing and the hosting of the
images. Carrierweave is the other part that handles the uploads and
model associations on our end.

This also builds out the admin interface. In order to keep things
simple, products have many images. This is to prevent weird JSON store
type things in Postgres and make future database editing simpler.