Fix WebConsole Whining
Webconsole is the plugin that provides a console embedded in the website
or as typically seen on error pages. It was throwing a whiny exception
about not being available on remote ports so I added the docker network
to the whitelist of domains it can run under.

This is documented in the README
https://github.com/rails/web-console