Large Refactor and Admin UI
This seperates the admin from the store using namespaces. Now everything
store related is in the store module and everything admin related is in
the admin module. Two seperate templates are also present with two
different sets of assets.

The reason for doing this is that the admin will stay the same between
clients and stores however the store could change drastically in the
front end requirements.

In addition there is quite a bit here to flush our the admin UI to the
point where I am finally decently happy with it.