Email Sign Up
Fix an error with the parameters in the email sign up. This also changes
the message that appears after the signup and adds a feature test around
it. One annoying thing was that the session needs to be set manually in
the session by the controller otherwise it does not work for some
reason. To circumvent this I put a before action on the application
controller that writes to the session.