Change Order Item Lookup
Originally skus that were shown on the product page had the cart skus
subtracted from them. This means there would be a good bit of additinal
validation and UI states that I do not want to consider supporting at
this time. I removed this subtraction to simplify the UI but then had to
change the way that items were added to the cart.

Now they get looked up via a find or create. If it finds it no
additional action is taken and the response returns opening the cart. If
it does not find it then it is created and adds it to the cart. This
ensures that duplicate items are not being added to the cart while a
user is shopping.