Finish Store Flow
This finishes the final pages for how the store will perform. The
homepage will load the categories. There is also a shop button that goes
to the category index for shop all products. Otherwise the category show
will load that category and the products for that category. The products
can then be loaded and shown in order to be purchased.