Fix Header On Mobile
I had the li options in the wrong list so they did not appear when the
button toggle was selected. I updated this, pulled the cart button out
to the other list and added a show class to ensure its always
displaying. Then floated the list right to make it appear like it did
before.