Update Images to allow Multiple Types
I needed to add a photo type field to the images table n order to support
distinct has_one with polymorphic associations. In order to support this
I had to carry this parameter into the form and the controller.
Admittedly this is now starting to get as complicated as I would it to
be and probably can't handle any further abstractions.